backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-generativeai
python-dotenv
pymupdf
python-docx
streamlit
requests
sqlalchemy
aiofiles
pillow
pytesseract
